Revert "Append single char as byte (where its known to fit) to avoid (small) allocation of String on each JSON object" Since `BufferEngine.print(Byte)` intended to print byte as number, not code point. Added test for quoted JSON values

+++ runtime/common/src/main/kotlin/kotlinx/serialization/json/JSON.kt
@@ -197,7 +197,7 @@ data class JSON(
         fun print(v: Boolean) = engine.print(v)
 
         fun printQuoted(value: String): Unit = with(engine) {
-            print(STRING.toByte())
+            print(STRING)
             var lastPos = 0
             val length = value.length
             for (i in 0 until length) {
@@ -211,7 +211,7 @@ data class JSON(
                 lastPos = i + 1
             }
             append(value, lastPos, length)
-            print(STRING.toByte())
+            print(STRING)
         }
     }
